glpk: move it to third_party_solvers/
