+++ ortools/math_opt/solvers/BUILD.bazel
@@ -442,8 +442,6 @@ cc_library(
         "//ortools/base:protoutil",
         "//ortools/base:status_macros",
         "//ortools/base:sysinfo",
-        "//ortools/glpk:glpk_env_deleter",
-        "//ortools/glpk:glpk_formatters",
         "//ortools/math_opt:callback_cc_proto",
         "//ortools/math_opt:infeasible_subsystem_cc_proto",
         "//ortools/math_opt:model_cc_proto",
@@ -464,6 +462,8 @@ cc_library(
         "//ortools/math_opt/solvers/glpk:rays",
         "//ortools/math_opt/validators:callback_validator",
         "//ortools/port:proto_utils",
+        "//ortools/third_party_solvers/glpk:glpk_env_deleter",
+        "//ortools/third_party_solvers/glpk:glpk_formatters",
         "//ortools/util:solve_interrupter",
         "@abseil-cpp//absl/cleanup",
         "@abseil-cpp//absl/container:flat_hash_map",
