+++ ortools/math_opt/solvers/glpk/BUILD.bazel
@@ -29,8 +29,8 @@ cc_library(
     deps = [
         "//ortools/base:logging",
         "//ortools/base:status_macros",
-        "//ortools/glpk:glpk_computational_form",
-        "//ortools/glpk:glpk_formatters",
+        "//ortools/third_party_solvers/glpk:glpk_computational_form",
+        "//ortools/third_party_solvers/glpk:glpk_formatters",
         "@abseil-cpp//absl/log:check",
         "@abseil-cpp//absl/status",
         "@abseil-cpp//absl/status:statusor",
