+++ ortools/math_opt/solvers/glpk/rays.cc
@@ -23,8 +23,8 @@
 #include "absl/strings/str_cat.h"
 #include "ortools/base/logging.h"
 #include "ortools/base/status_macros.h"
-#include "ortools/glpk/glpk_computational_form.h"
-#include "ortools/glpk/glpk_formatters.h"
+#include "ortools/third_party_solvers/glpk/glpk_computational_form.h"
+#include "ortools/third_party_solvers/glpk/glpk_formatters.h"
 
 namespace operations_research::math_opt {
 namespace {
