+++ ortools/third_party_solvers/glpk/glpk_formatters.h
@@ -12,8 +12,8 @@
 // limitations under the License.
 
 // Formatting functions for GLPK constants.
-#ifndef ORTOOLS_GLPK_GLPK_FORMATTERS_H_
-#define ORTOOLS_GLPK_GLPK_FORMATTERS_H_
+#ifndef ORTOOLS_THIRD_PARTY_SOLVERS_GLPK_GLPK_FORMATTERS_H_
+#define ORTOOLS_THIRD_PARTY_SOLVERS_GLPK_GLPK_FORMATTERS_H_
 
 #include <cstddef>
 #include <string>
@@ -41,4 +41,4 @@ std::string TruncateAndQuoteGLPKName(std::string_view original_name);
 
 }  // namespace operations_research
 
-#endif  // ORTOOLS_GLPK_GLPK_FORMATTERS_H_
+#endif  // ORTOOLS_THIRD_PARTY_SOLVERS_GLPK_GLPK_FORMATTERS_H_
